% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymization.R
\name{anonymization}
\alias{anonymization}
\title{Anonymize a snapshot of a dynamic network.}
\usage{
anonymization(g, alpha = 1, beta = 2, gamma = 3)
}
\arguments{
\item{g}{A network grouped by lw-grouping algorithm.}

\item{alpha}{Weight of anonymization cost resulted from label generalization.}

\item{beta}{Weight of anonymization cost resulted from adding edges.}

\item{gamma}{Weight of anonymization cost resulted from adding nodes.}
}
\description{
Anonymize a snapshot of a dynamic network.
}

