% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymize2node.R
\name{anonymize2node}
\alias{anonymize2node}
\title{Anonymize two node.}
\usage{
anonymize2node(g, uid, vid, noise = g$noise)
}
\arguments{
\item{g}{A graph contains vertexs with different labels and some of which are sensitive.}

\item{uid}{Name of a node with sensitive label.}

\item{vid}{Name of a node with unsensitive label.}

\item{noise}{Current amount of noise nodes.}
}
\value{
A list with information of anonymized network.
}
\description{
Anonymize two node.
}

