% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost.R
\name{cost}
\alias{cost}
\title{Calculate anonymization cost of two nodes.}
\usage{
cost(g, uid, vid, alpha = 1, beta = 2, gamma = 3)
}
\arguments{
\item{g}{A graph contains vertexs with different labels and some of which are sensitive.}

\item{uid}{Name of a node with sensitive label.}

\item{vid}{Name of a node with unsensitive label.}

\item{alpha}{Weight of anonymization cost resulted from label generalization.}

\item{beta}{Weight of anonymization cost resulted from adding edges.}

\item{gamma}{Weight of anonymization cost resulted from adding nodes.}
}
\value{
Anonymization cost of two nodes.
}
\description{
Calculate anonymization cost of two nodes.
}

