% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lw_grouping.R
\name{lw.grouping}
\alias{lw.grouping}
\title{Generate a grouped dynamic network by lw-grouping algorithm.}
\usage{
lw.grouping(dynet = NULL, l = 2, w = 3)
}
\arguments{
\item{dynet}{An ungrouped dynamic network.}

\item{l}{Kinds of labels in each unmerged group.}

\item{w}{Width of window of lw-grouping algorithm.}
}
\value{
A list of grouped network with attribute of gs.merged.
}
\description{
Generate a grouped dynamic network by lw-grouping algorithm.
}

