% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_virtual_dynamic_network.R
\name{make.virtual.dynamic.network}
\alias{make.virtual.dynamic.network}
\title{Make a vertex-increasing virtual dynamic network.}
\usage{
make.virtual.dynamic.network(network.data = NULL, len = 10, by = 5,
  label.types = 100, prop.init = 0.001, prop.sensitive = 0.1)
}
\arguments{
\item{network.data}{A data frame containing a symbolic edge list,which contains the information of whole network data.}

\item{len}{Time of this dynamic network lasts.}

\item{by}{The number of vertex added in network each time.}

\item{label.types}{The number of label types the network possesses.}

\item{prop.init}{The proportion of vertex amounts of initial network in whole network data.}

\item{prop.sensitive}{The proportion of amounts of vertex with sensitive label in whole network data.}
}
\value{
A list of snapshots of a virtual dynamic network.
}
\description{
Make a vertex-increasing virtual dynamic network.
}
\examples{
dynet <- make.virtual.dynamic.network()
}

