% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\docType{data}
\name{network}
\alias{network}
\title{Unirected graph: CA-CondMat}
\format{An object of class \code{data.frame} with 93439 rows and 2 columns.}
\usage{
network
}
\description{
Collaboration network of Arxiv Condensed Matter category (there is an edge if authors coauthored at least one paper)
network
}
\details{
@format A data frame with 93439 rows and 2 variables:
}
\keyword{datasets}

